requirements.txt
Cython==0.29.23
numpy==1.19.5
opencv-python==4.5.3.56
pypotrace==0.3
